DRY payment scheduler

+++ mortgage/api/payment_schedule.py
@@ -1,6 +1,6 @@
-from math import pow
 from decimal import Decimal
 from dateutil import relativedelta
+from math import pow
 
 from mortgage.models import Payment
 
@@ -11,10 +11,9 @@ class PaymentScheduler:
         self.extra_payment = extra_payment
 
     def calc_payments_schedule(self, start_payment_number=1, amount=None):
-        monthly_percent = Decimal(self.mortgage.percent / (12 * 100))  # 1/12 of credit's percent in 0.xx format
-        power = Decimal(pow(monthly_percent + 1, self.mortgage.period_in_months))  # 1 - hardcode in math formula
-        coef = Decimal(power * monthly_percent / (power - 1))
-
+        # 1 - hardcode in math formula
+        power = Decimal(pow(self.mortgage.monthly_percent + 1, self.mortgage.period_in_months))
+        coef = Decimal(power * self.mortgage.monthly_percent / (power - 1))
         amount = round(coef * self.mortgage.total_amount, 2) if not amount else amount
 
         for i in range(start_payment_number, self.mortgage.period_in_months + 1):
@@ -51,28 +50,20 @@ class PaymentScheduler:
         extra_payment.debt_rest = extra_payment.calc_debt_rest()
         extra_payment.save()
 
-        # TODO: if there is no next payment?
         next_payment = self.extra_payment.get_next_payment()
-        time_between_payments = next_payment.date - extra_payment.date
-        days_between_payments = time_between_payments.days
-        next_payment.amount = extra_payment.debt_rest * self.mortgage.percent / 100 / 365 * days_between_payments
-        next_payment.bank_amount = next_payment.amount
-        next_payment.debt_decrease = 0
-        next_payment.debt_rest = next_payment.calc_debt_rest()
-        next_payment.save()
+        if next_payment:
+            time_between_payments = next_payment.date - extra_payment.date
+            days_between_payments = time_between_payments.days
+            next_payment.amount = extra_payment.debt_rest * self.mortgage.percent / 100 / 365 * days_between_payments
+            next_payment.bank_amount = next_payment.amount
+            next_payment.debt_decrease = 0
+            next_payment.debt_rest = next_payment.calc_debt_rest() - extra_payment.debt_decrease
+            next_payment.save()
 
-        time_left = relativedelta.relativedelta(self.mortgage.last_payment_date, next_payment.date)
-        # count of payments between extra payment and last payment
-        months_left = time_left.months + time_left.years * 12
-        monthly_percent = Decimal(self.mortgage.percent / (12 * 100))  # 1/12 of credit's percent in 0.xx format
-        power = Decimal(pow(monthly_percent + 1, months_left))  # 1 - hardcode in math formula
-        coef = Decimal(power * monthly_percent / (power - 1))
+            Payment.objects.filter(mortgage_id=self.mortgage.pk, date__gt=next_payment.date).delete()
 
-        Payment.objects.filter(mortgage_id=self.mortgage.pk, date__gt=next_payment.date).delete()
-
-        start_payment_number = relativedelta.relativedelta(next_payment.date, self.mortgage.issue_date).months + 1
-        amount = round(coef * extra_payment.debt_rest, 2)
-        self.calc_payments_schedule(start_payment_number, amount)
+            start_payment_number, amount = self.get_new_schedule_parameters(next_payment)
+            self.calc_payments_schedule(start_payment_number, amount)
 
     def less_than_bank_percent_saving(self):
         extra_payment = self.extra_payment
@@ -81,13 +72,13 @@ class PaymentScheduler:
         extra_payment.debt_rest = extra_payment.calc_debt_rest()
         extra_payment.save()
 
-        # TODO: if there is no next payment?
         next_payment = extra_payment.get_next_payment()
-        next_payment.amount = next_payment.amount - self.extra_payment['amount']
-        next_payment.bank_amount = next_payment.calc_bank_amount()
-        next_payment.debt_decrease = next_payment.calc_debt_decrease()
-        next_payment.debt_rest = next_payment.calc_debt_rest()
-        next_payment.save()
+        if next_payment:
+            next_payment.amount = next_payment.amount - self.extra_payment['amount']
+            next_payment.bank_amount = next_payment.calc_bank_amount()
+            next_payment.debt_decrease = next_payment.calc_debt_decrease()
+            next_payment.debt_rest = next_payment.calc_debt_rest()
+            next_payment.save()
 
     def more_than_bank_percent_saving(self):
         extra_payment = self.extra_payment
@@ -96,23 +87,26 @@ class PaymentScheduler:
         extra_payment.debt_rest = extra_payment.calc_debt_rest()
         extra_payment.save()
 
-        # TODO: if there is no next payment?
         next_payment = extra_payment.get_next_payment()
-        next_payment.bank_amount = next_payment.calc_bank_amount()
-        next_payment.amount = next_payment.bank_amount
-        next_payment.debt_decrease = next_payment.calc_debt_decrease()
-        next_payment.debt_rest = next_payment.calc_debt_rest()
-        next_payment.save()
+        if next_payment:
+            next_payment.bank_amount = next_payment.calc_bank_amount()
+            next_payment.amount = next_payment.bank_amount
+            next_payment.debt_decrease = next_payment.calc_debt_decrease()
+            next_payment.debt_rest = next_payment.calc_debt_rest()
+            next_payment.save()
 
-        Payment.objects.filter(mortgage_id=self.mortgage.pk, date__gt=next_payment.date).delete()
+            Payment.objects.filter(mortgage_id=self.mortgage.pk, date__gt=next_payment.date).delete()
 
+            start_payment_number, amount = self.get_new_schedule_parameters(next_payment)
+            self.calc_payments_schedule(start_payment_number, amount)
+
+    def get_new_schedule_parameters(self, next_payment):
         time_left = relativedelta.relativedelta(self.mortgage.last_payment_date, next_payment.date)
         # count of payments between extra payment and last payment
         months_left = time_left.months + time_left.years * 12
-        monthly_percent = Decimal(self.mortgage.percent / (12 * 100))  # 1/12 of credit's percent in 0.xx format
-        power = Decimal(pow(monthly_percent + 1, months_left))  # 1 - hardcode in math formula
-        coef = Decimal(power * monthly_percent / (power - 1))
+        power = Decimal(pow(self.mortgage.monthly_percent + 1, months_left))
+        coef = Decimal(power * self.mortgage.monthly_percent / (power - 1))
 
         start_payment_number = relativedelta.relativedelta(next_payment.date, self.mortgage.issue_date).months + 1
-        amount = round(coef * next_payment.debt_rest, 2)
-        self.calc_payments_schedule(start_payment_number, amount)
+        amount = round(coef * self.extra_payment.debt_rest, 2)
+        return start_payment_number, amount
